Put govuk_elements assets at start of asset pipeline load path
The SASS in the govuk_elements[1] bower component refers to
/assets/icons/icon-pointer.png when compiled but the same file is also provided
by the govuk_frontend_toolkit[2] gem at /assets/icon-pointer.png which is what
ends up in the compiled assets.

Put the bower component asset path before those of the gem so the correct url
is used in the compiled stylesheet.

[1] https://github.com/alphagov/govuk_elements/blob/master/govuk/public/images/icons/icon-pointer.png
[2] https://github.com/alphagov/govuk_frontend_toolkit/blob/master/images/icon-pointer.png