Improvements to guaranteed income guide
Changed title to fit in with how it appears in the navigation. We went with 'Get' rather than 'Getting' to make it easier to read in the navigation - a list with 'ing' on every word at the front would make it difficult to navigate.
Added ##Overview heading for the right-hand page navigation
Changed the opening sentences to bullet points, giving an overview of the option because user research showed that people wanted an 'idiot's guide' that was even simpler than the page content.
Shortened headings to make the page easier to navigate and give users the keywords they are looking for.
Moved text around the page to make it flow better.
Rewrote some text to make it simpler.
Took out superfluous text - including lead-in sentences to bullets - to make the page easier to read.
Changed 'charges' to 'fees' because keyword research shows that 'charges' is a word that users don't understand.
Added 'Book an appointment' sentence and link because that is a next step. Analytics show that users aren't clicking on the 'Book an appointment' link in the right-hand navigation at the bottom of the page. We will use Analytics to see whether this works or whether the link that is being added to the top right of the pages works.
Changed link to FCA to one of our pages because we should send users to our pages to explain it.