Simplify information about pension recycling
Changed pension recycling information in line with new fact checks from HMRC. Took out factually incorrect information. Shortened the information to just the basics because the detail of the information is too confusing to understand on a web page. Guidance is to see a financial adviser if you're thinking of paying in money to your pension at the same time as taking it out.