Create new pension recycling page
Moved from the tax page after 2 rounds of lab testing. Users were confused by it and it took a lot of attention away from the rest of the tax page.

Dropped the 55% tax figure because users were fixating on that, getting irate about it, and not taking in the more important information about when they could be breaching pension recycling law.