Create Booking Request via API
Completes the basic success case for booking requests. At the moment
this does not checks for timeouts, exceptions, error handling or
validation. This simply completes the success-case for this particular
story.

Had to add a wait between the last remaining slot being chosen and
the JS doing its thing before clicking the `Complete` button on the
calendar view. This was flakey locally due to a timing issue, and
almost always failed in slower environments - ie travis-ci.