Cache booking locations in memory
Use a read-through cache to store booking locations JSON in memory and
save a bunch of roundtrips to the API during the booking process. The
locations have a TTL of two hours by default.

For the timebeing we're configured to use the memory store cache, but
it may make more sense in future to use the redis cache or something
shared, not per-dyno.