Simplify the guide implementation
This change is designed to simplify the process of
adding metadata to a guide page.

Previously all fields needed to be declared in
multiple places and explicitly initialised.

This change is a small improvement to the existing
codebase with the goal of simplifying the current
process. This change has been scope limited and is
not intended to fix all the outstanding issues.