Enable CSRF protection where possible
After upgrading to Rails 5.0.1 CSRF protection attempts to clear the
flash upon failure, unfortunately a bug in Rails means it doesn't guard
against us not actually loading the session middleware thus causing the
app to crash where usually CSRF failures would be silent.

We now fail with exceptions for CSRF failures so we can catch them
early on. For the bookings, where we take sensitive data we absolutely
should ensure CSRF verification happens so I have enabled this. For the
other tools this isn't such an issue.

Given the nature of the pension type tool we cannot enable CSRF
verification so this is disabled entirely.