Enable multiple locales
The locale needs to be set on a per request basis
passed in as one of the parameters.

Guides are now aware of `available_locales` that they
can be rendered in, and will be rendered in using the
current `I18n.locale` with a fallback to `:en`.

Each local needs a language translation in a
`config/locales/<locale>.yml` file.

Guides inherit metadata from base file.

Creating a new guide for for a location is as simple
as creating a copy of the guide file with a `.<locale>.md`
extension.

No front-end implementation is included in this PR.

It is expect the front-end code will be similar to:

```erb
<% if @guide.available_locales.count > 1 %>
  <% @guide.available_locales.each do |locale| %>
    <% if locale == I8n.locale %>
      <span><%= I18n.t(:language, locale: locale) %></span>
    <% else %>
     <%= link_to I18n.t(:language, locale: locale || :en), public_send("#{@guide.id}_path", locale: locale) %>
    <% end %>
  <% end %>
<% end %>
```