Expose an API for clearing booking locations
We need a way of remotely triggering a cache flush when booking
locations change structurely. This change adds an API endpoint, with
simple token based authentication. We are not using SSO here since this
application is not backed by a persistent store - a requirement for
GDS-SSO and signon integration.

A request issued to `DELETE /locations_cache` with the correct
authorisation token will sweep the cached locations and leave
non-prefixed (anything other than our locations) entries as-is.