Alter wording on location page to better clarify phone booking hours
Customers have been geting confused about the wording of the
booking hours on the location pages. They were thinking that the
hours specified meant when the location was available for an
appointment rather than when the phone lines were open.

This alters the wording slightly to help reduce this confusion.