Update robots.txt to take into account welsh language URLs
This puts a wildcard language directory at the start of URLs
in the robots.txt file.

It also removes some URLs from the robots.txt which are no
longer valid, and adds in /*/booking-request/ to avoid
step one of the f2f booking process from being indexed.

It was felt that allowing customers to go straight to the
booking form without explaining who is suitable for a
Pension Wise appointment might increase the chances of
getting invalid appointments.