Adding an additional info field to F2F booking
This introduces an additional information text field which allows
customers to specify an additional amount of information about
their booking request e.g. times they aren't available during
a AM/PM slot.

It also introduces a character limit module in JS which:

* Limits input to the textbox to 160 chars on older browsers
which don't support the maxlength attribute on a `<textarea>`
* Displays a character remaining count underneath the textarea
* Makes use of the aria-live and aria-atomic attributes when
the user is getting low on characters left, so screen reader
users are aware that they are starting to hit the character limit