Don't error when refreshing a confirmation
When customers return to an old browser session their device may refresh
the confirmation page without the requisite parameters. We are seeing
this occur a few times weekly in our error tracker. Rather than display
an error we should simply return them to the homepage.

I've only checked the presence of the `booking_reference` parameter
since one will not exist without the other when refreshing.