Don't show phone number with no availability
When the location is enabled for online bookings and the location has
no availability the customer was always presented the call centre
number. If they called the number they would be told the location has
no availability and they would reach a dead end. This change ensures
they are made aware early on when the location has no appointment
availability.