Remove the stickiness of the explore-your-options/summary sidebar
Prior to this - the sidebar on /en/explore-your-options/summary disappears in the mobile view rather than be positioned underneath the main content. This is bad because the sidebar contains buttons the user needs to be able to click.

Looking around the app - /en/summary-document is another page using the same "sticky_sidebar" module, where the sidebar also disappears in the mobile view. However, here it's likely that's intended behaviour since the sidebar just contains subsection navigation.

The `.sticky-sidebar` CSS module is hidden by default (the mobile view), and becomes visible for tablets and above.

Stickiness aside /en/explore-your-options/summary is making use of `.sticky-sidebar__nav`, `.sticky-sidebar__heading`, `sticky-sidebar__list` etc. I am guessing we want to keep it's current visual appearance while getting rid of it's disappearing/sticky trick.

This commit probably makes it appear as intended... but it probably makes sense to extract the relevant styles into a separate module.