Disable mailgun validations at will
Due to the fact we sometimes hit the limit for mailgun validation API
requests and their inability to allow us to purchase more, since the
service was provisioned via heroku - we needed a way to quickly disable
customer-facing validations if necessary.

The presence of the `DISABLE_MAILGUN_VALIDATION` environment variable
will disable customer validations. Unsetting the variable will switch
them back on.