Ensure appointment confirmations are displayed
Previously we were checking an incorrect predicate to determine whether
the booking was automatically fulfilled in realtime or not. Also, this
was fragile when a customer returned to the confirmation page by
refresh and the required `flash` data was no longer present. This
approach ensures bookings and appointments are displayed correctly
without error.