Issue 29: Validation and form help text
After RNIB feedback the wording for the phone prompt is clearer and the
required fields throughout booking journeys are marked as required. This
will aid screen readers and non screen readers alike as the requirements
will be announced for the former and forms will not be submitted for
both kinds of users when they are not completed.