Issues 48, 52, 56, 60 - display errors consistently
Makes an attempt at styling and displaying errors more inline with how
they're displayed elsewhere. Due to how these question/answer pages are
constructed it was not possible to add an error context that can be used
to predicate upon un the view/markdown. Instead, I used javascript to
pluck what I needed from the error summary and add the correct classes
and labelling where necessary.

In future, we will move away from the guides approach and convert this
to a fully-fledged journey.