Switch to chrome driver for JS feature tests
Being pinned to older versions of site_prism and capybara was holding us
back so since the modern approach to feature tests is with webdriver and
chrome we ought to use this instead.

Govspeak was bumped for compatibility reasons as the version we had
pinned was relying on an outdated version of addressable that
transitively blocked upgrades to other gems.