Add a provider embed for nudge appointments
When provided with the embed URL, a pension provider can now embed our
nudge journey for the customers. Two cookies are assigned on the
'landing page' to signify that this is a nudge booking and that it has
been embedded thus the 'where did you hear' ought to be set to '23'
signifying the 'External (Stronger Nudge)' WDYH in the planner.