Adjust a11y requirements behaviours
This adds an extra field for capturing a11y adjustments, breaking it out
from the existing notes field that was used previously. This is to
ensure that when the customer checks the a11y requirements box, we can
be sure that the notes they provide relate to their adjustments and not
mixed or interspersed with other additional information they provide.