Adjust a11y requirements behaviours (#1538)
* Adjust a11y requirements behaviours

This adds an extra field for capturing a11y adjustments, breaking it out
from the existing notes field that was used previously. This is to
ensure that when the customer checks the a11y requirements box, we can
be sure that the notes they provide relate to their adjustments and not
mixed or interspersed with other additional information they provide.

* Tidy up remaining form elements

* Alterations to hints for additional notes fields

* Add link to adjustments details

* Don't display the form hint for PSG