Upgrade to Rails 8 and latest Ruby (#1547)
* Bump Rails to 7

* Remove date formatting deprecation warnings

* Compatibility fixes

* Change the signature for verified messages

This has changed slightly so whilst existing, shared messages will be
invalid, they are ephemeral so customers will not be affected.

* Remove this redundant path

* Bump to Ruby 3.4

Also bumps required gems for compatibility

* Switch away from secrets to credentials

This is only used for the `SECRET_KEY_BASE` but should be used for all
environment's secrets going forward.

* Bump to Rails 8

* Correct ordering for rubocop

* Remove legacy redis stores

These are replaced with built in redis cache stores so can be removed.

* Explicitly require the API class

* Correct uglifier configuration

* Correct paths for loader

* Converting value to an integer

This causes errors deeper in Rails when it's used for `expires_in`
during cache reads/writes.